webserver/requirements.txt
panns-inference==0.0.7
librosa==0.8.0
numpy==1.16.5
pymilvus==0.3.0
pillow
pymysql
fastapi
uvicorn
pandas